use mixin for common methods in views
use initial method in new mixin and use new F680s case filter to narrow queryset

use initial method in mixin and use new filter backend
