+++ api/f680/caseworker/filters.py
@@ -1,6 +1,12 @@
 from rest_framework import filters
+from api.cases.enums import CaseTypeSubTypeEnum
 
 
 class CurrentCaseFilter(filters.BaseFilterBackend):
     def filter_queryset(self, request, queryset, view):
         return queryset.filter(case_id=view.kwargs["pk"])
+
+
+class F680CaseFilter(filters.BaseFilterBackend):
+    def filter_queryset(self, request, queryset, view):
+        return queryset.filter(case__case_type__sub_type=CaseTypeSubTypeEnum.F680)
