+++ api/f680/caseworker/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from api.f680.caseworker.views import F680RecommendationViewSet, F680OutcomeViewSet, F680OutcomeDocumenViewSet
+from api.f680.caseworker.views import F680RecommendationViewSet, F680OutcomeViewSet, F680OutcomeDocumentViewSet
 
 app_name = "caseworker_f680"
 
@@ -37,7 +37,7 @@ urlpatterns = [
     ),
     path(
         "<uuid:pk>/outcome_document/",
-        F680OutcomeDocumenViewSet.as_view(
+        F680OutcomeDocumentViewSet.as_view(
             {
                 "get": "list",
             }
