+++ api/f680/caseworker/views.py
@@ -1,43 +1,27 @@
-from django.core.exceptions import ObjectDoesNotExist
-from django.http import Http404, HttpResponse
+from django.http import HttpResponse
 from api.cases.generated_documents.models import GeneratedCaseDocument
 from rest_framework import status, viewsets
 from rest_framework.response import Response
 
-from api.applications.libraries.get_applications import get_application
 from api.audit_trail import service as audit_trail_service
 from api.audit_trail.enums import AuditType
-from api.core.authentication import GovAuthentication
-from api.core.exceptions import NotFoundError
 
 from api.f680.models import Recommendation, SecurityReleaseOutcome
-from api.f680.caseworker import filters
 from api.f680.caseworker import permissions
 from api.f680.caseworker import serializers
 from api.f680.caseworker import read_only_serializers
 
+from .mixins import F680CaseworkerApplicationMixin
 
-class F680RecommendationViewSet(viewsets.ModelViewSet):
-    authentication_classes = (GovAuthentication,)
+
+class F680RecommendationViewSet(F680CaseworkerApplicationMixin, viewsets.ModelViewSet):
     permission_classes = [
         permissions.CaseCanAcceptRecommendations & permissions.CaseCanUserMakeRecommendations | permissions.ReadOnly
     ]
-    filter_backends = (filters.CurrentCaseFilter,)
+
     queryset = Recommendation.objects.all()
     serializer_class = serializers.F680RecommendationSerializer
     pagination_class = None
-
-    def dispatch(self, request, *args, **kwargs):
-        # TODO: Review dispatch methods in LTD-6085
-        try:
-            self.application = get_application(self.kwargs["pk"])
-        except (ObjectDoesNotExist, NotFoundError):
-            raise Http404()
-
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_case(self):
-        return self.application
 
     def prepare_data(self, request_data):
         return [
@@ -96,26 +80,12 @@ class F680RecommendationViewSet(viewsets.ModelViewSet):
         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
 
 
-class F680OutcomeViewSet(viewsets.ModelViewSet):
-    authentication_classes = (GovAuthentication,)
+class F680OutcomeViewSet(F680CaseworkerApplicationMixin, viewsets.ModelViewSet):
     permission_classes = [permissions.CaseReadyForOutcome & permissions.CanUserMakeOutcome | permissions.ReadOnly]
-    filter_backends = (filters.CurrentCaseFilter,)
     queryset = SecurityReleaseOutcome.objects.all()
     serializer_class = serializers.SecurityReleaseOutcomeSerializer
     lookup_url_kwarg = "outcome_id"
     pagination_class = None
-
-    def dispatch(self, request, *args, **kwargs):
-        # TODO: Review dispatch methods in LTD-6085
-        try:
-            self.application = get_application(self.kwargs["pk"])
-        except (ObjectDoesNotExist, NotFoundError):
-            raise Http404()
-
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_case(self):
-        return self.application
 
     def prepare_data(self, request_data):
         return {
@@ -165,18 +135,8 @@ class F680OutcomeViewSet(viewsets.ModelViewSet):
         return HttpResponse(status=response.status_code)
 
 
-class F680OutcomeDocumenViewSet(viewsets.ModelViewSet):
-    authentication_classes = (GovAuthentication,)
-    filter_backends = (filters.CurrentCaseFilter,)
+class F680OutcomeDocumentViewSet(F680CaseworkerApplicationMixin, viewsets.ModelViewSet):
     queryset = GeneratedCaseDocument.objects.all()
     serializer_class = serializers.OutcomeDocumentSerializer
     lookup_url_kwarg = "case_id"
     pagination_class = None
-
-    def dispatch(self, request, *args, **kwargs):
-        try:
-            self.application = get_application(self.kwargs["pk"])
-        except (ObjectDoesNotExist, NotFoundError):
-            raise Http404()
-
-        return super().dispatch(request, *args, **kwargs)
